dspqueue: drop dspqueue_shared.h
The dspqueue_sahred.h header is only used by dspqueue_cpu.c. Inline
header contents into the corresponding file.
